Convert "template" to subcommand autodetection

+++ src/lxq-template-destroy.sh
@@ -1,24 +1,10 @@
 #!/usr/bin/env bash
-
-# This script is based on the template here:
-#
-#     https://gist.github.com/pcrockett/8e04641f8473081c3a93de744873f787
-#
-# It was copy/pasted here into this file and then modified extensively.
-#
-# Useful links when writing a script:
-#
-# Shellcheck: https://github.com/koalaman/shellcheck
-# vscode-shellcheck: https://github.com/timonwong/vscode-shellcheck
-#
-# I stole many of my ideas here from:
-#
-# https://blog.yossarian.net/2020/01/23/Anybody-can-write-good-bash-with-a-little-effort
-# https://dave.autonoma.ca/blog/2019/05/22/typesetting-markdown-part-1/
-#
-
-# https://vaneyckt.io/posts/safer_bash_scripts_with_set_euxo_pipefail/
 set -Eeuo pipefail
+
+if is_set "${LXQ_SHORT_SUMMARY+x}"; then
+    printf "\t\tDestroy a template"
+    exit 0
+fi
 
 readonly DEPENDENCIES=(lxc-destroy)
 readonly TEMPLATES_CONFIG_DIR="${LXQ_REPO_DIR}/templates"
